requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.7
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7
